RM 1215-2356 (on 2399): Basemap updates
- Added a new version of the basemap config file that can handle a mix
of perspective and cylindrically projected images, as well as new
metadata to handle descriptions (this was needed for MEGANE)
- Updated the SmallBodyControllPanel to handle both the v1 and v2
versions; v2 can display an Info button to display the basemap
description
- Performance improvements for vtkImageRendererOperator (pushing some
operations to background thread)
- Cylindrical images can now fetch both pre-generated polydata and
VTKImageData (which for large files can take a while to load; loading
the saved data is ~50-60% faster)
- This still uses the legacy Custom key classes, but has started to
introduced deprecations into the methods that use them; once all server
config files have been moved to the v2 basemap config data, those types
will be retired once and for all, and the code (e.g.
SmallBodyControlPanel and others) will be cleaned up.
- This is isolated enough so it can be brought over to 1215-2356, since
it is heavily image reorg related (it was something that started to get
implemented but got put off; this was a good way to implement it with
some good test cases)